updated exercise

+++ albums_migration.sql
@@ -7,7 +7,10 @@ id INT UNSIGNED NOT NULL AUTO_INCREMENT,
 artist VARCHAR(100),
 name VARCHAR(100),
 release_date INT,
-sales FLOAT,
+sales DECIMAL(6,2),
 genre VARCHAR(100),
 PRIMARY KEY (id)
 );
+
+
+
